* 源端/租户端批量拉全量功能添加 * log类型数据源表级别规则组导入导出功能添加 * 优化项目删除逻辑 * 拉全量历史记录添加id,targetSinkTopic查询条件,orderby排序条件前端页面开发 * 批量发送ctrlMessage功能添加 * mysql5.5版本初始化meta version information_schema.COLUMNS缺少DATETIME_PRECISION列的bug修复 * 心跳插入线程拆分 * kafka小工具限制返回值大小1m * 全量无异常重启bug修复(jdbc查询fetchSize设置为0导致) * oracle加表前先校验源端全量补充日志是否打开(此功能需要开通dba_log_groups的访问权限) * 生成OGG Trail前缀名称功能添加 * 修改源端和租户端加表后,表默认为启动状态 * 项目表页面添加批量停止,批量启动, 批量删除功能 * 数据源管理页面添加ogg或canal的信息展示 * 租户端添加条件拉全量功能 * ogg replicat,canal自动化部署添加功能添加(包含自动添加删除replicat进程,自动添加删除ogg抽取表配置;canal自动化部署启动,停止) * 修改表结构变更后，找不到version导致空指针异常的bug * 为了避免dbus-router统计信息与dispatcher和appender不一致的情况，将写心跳到kafka中的逻辑提前到appender-kafka-writer中 * 项目表添加后默认拥有拉全量权限 * 添加项目级脱敏和项目表脱敏时，显示该列是否有DBA脱敏和项目级脱敏，并使用红色的锁图标标记 * 取消influxdb内网正则表达式校验(针对首次初始化点击提交按钮没有反应的bug) * log修复save as算子未对空字符串处理的bug * t_project_topo_table_encode_output_columns中增加特批不脱敏special_approve字段 * 线路检查bug修改

+++ dbus-keeper/keeper-base/src/main/java/com/creditease/dbus/domain/model/ProjectTopoTableMetaVersion.java
@@ -37,7 +37,7 @@ public class ProjectTopoTableMetaVersion {
 
     private String dataType;
 
-    private Integer dataLength;
+    private Long dataLength;
 
     private Date updateTime;
 
@@ -113,11 +113,11 @@ public class ProjectTopoTableMetaVersion {
         this.dataType = dataType;
     }
 
-    public Integer getDataLength() {
+    public Long getDataLength() {
         return dataLength;
     }
 
-    public void setDataLength(Integer dataLength) {
+    public void setDataLength(Long dataLength) {
         this.dataLength = dataLength;
     }
